Transcription feedback during recording.
Moved audio recording to a separate thread, freeing up the main thread
for UI updates. Note that record_melody() and record_rhythm() are now
non-blocking as a result, which is pretty scary. We should probably
figure out the Right Way to do it later. (async/await?)

Also, we should touch up the UI itself a bit, and make sure that the
recording thing is always in the center of the screen. (Not a particular
position on the canvas.)