At long last, playback on Android.
After much debugging, I realized the actual issue was that the soundfont
wasn't being included in the packaged app. D'oh.

Right now audio is going from fluidsynth to audiostream, but this might
be unnecessary - I tried this because I was getting no sound output with
the built-in fluidsynth android drivers (opensles and oboe), but of
course the immediate issue there was the lack of SoundFont.