Basic chord support.
Just C, F, G, Am for demonstration.

Needs rhythm notation to improve usability; currently requires
triggering a note in addition to the chord.

(Actually, with normal notes, it's not clear that the chord should be
retriggered. Consider lead sheets, for example; chord marking indicates
'comp until next chord' (while playing the melody as written, not as an
indicator of chord rhythm).)