Memoization is easy
And performance is better (from ~1 second to ~.35).
Memoizes StrokeGroup recognition results, using frozen sets of the ids
of the enclosed strokes as key. This is safe because a stroke will never
be mutated after it is completed (unlike the old Gestures), and stroke
ids are not reused.

Note that this gets the number of recognize calls as low as it can go,
for re-interpreting the transcription ink. Further optimizations would
either have to continue speeding up recognition (perhaps by going full
$Q and implementing the LUTs) or by simply refusing the interpret the
transcription ink until necessary (e.g. until it is modified by
overlapping user ink).